Implementing RTC HAL: Added RTC HAL files to the project Implemented RTC APIs Modified existing code to call the RTC APIs instead of Driver APIs Moved Clock, PLL and Watchdog Timer initialization code under RTC Corrected the comments for SYSTEM_CLOCK_FREQ and USEC_PER_TICK
