+++ embedded/projects/osp-lpc54102/Keil/osp-lpc54102.uvprojx
@@ -1043,6 +1043,11 @@
               <FileType>1</FileType>
               <FilePath>..\..\..\..\targets\TARGET_LPC54102\pinmap.c</FilePath>
             </File>
+            <File>
+              <FileName>rtc_api.c</FileName>
+              <FileType>1</FileType>
+              <FilePath>..\..\..\..\targets\TARGET_LPC54102\rtc_api.c</FilePath>
+            </File>
           </Files>
         </Group>
         <Group>
